chore: reorganize and remove unused files
